Make it work on Windows
