+++ x/examples/fyne-tools/main.go
@@ -14,22 +14,12 @@
 
 package main
 
-/*
-#include <stdlib.h>
-#include <sys/types.h>
-#include <sys/socket.h>
-#include <netdb.h>
-*/
-import "C"
-
 import (
 	"context"
-	"fmt"
 	"image/color"
 	"log"
 	"net"
 	"strings"
-	"unsafe"
 
 	"fyne.io/fyne/v2"
 	"fyne.io/fyne/v2/app"
@@ -39,48 +29,6 @@ import (
 	"fyne.io/fyne/v2/theme"
 	"fyne.io/fyne/v2/widget"
 )
-
-func lookupCNAME(ctx context.Context, domain string) (string, error) {
-	type result struct {
-		cname string
-		err   error
-	}
-
-	results := make(chan result)
-	go func() {
-		cname, err := lookupCNAMEBlocking(domain)
-		results <- result{cname, err}
-	}()
-
-	select {
-	case r := <-results:
-		return r.cname, r.err
-	case <-ctx.Done():
-		return "", ctx.Err()
-	}
-}
-
-func lookupCNAMEBlocking(host string) (string, error) {
-	var hints C.struct_addrinfo
-	var result *C.struct_addrinfo
-
-	chost := C.CString(host)
-	defer C.free(unsafe.Pointer(chost))
-
-	hints.ai_family = C.AF_UNSPEC
-	hints.ai_flags = C.AI_CANONNAME
-
-	// Call getaddrinfo
-	res := C.getaddrinfo(chost, nil, &hints, &result)
-	if res != 0 {
-		return "", fmt.Errorf("getaddrinfo error: %s", C.GoString(C.gai_strerror(res)))
-	}
-	defer C.freeaddrinfo(result)
-
-	// Extract canonical name
-	cname := C.GoString(result.ai_canonname)
-	return cname, nil
-}
 
 type appTheme struct {
 	fyne.Theme
